mm/memory_hotplug.c: add a bounds check to check_hotplug_memory_range()
Patch series "Add bounds check for Hotplugged memory", v3.

This series adds bounds checks for hotplugged memory, ensuring that it is
within the physically addressable range (for platforms that define
MAX_(POSSIBLE_)PHYSMEM_BITS.

This allows for early failure, rather than attempting to access bogus
section numbers.


This patch (of 2):

On PowerPC, the address ranges allocated to OpenCAPI LPC memory are
allocated from firmware.  These address ranges may be higher than what
older kernels permit, as we increased the maximum permissable address in
commit 4ffe713b7587 ("powerpc/mm: Increase the max addressable memory to
2PB").  It is possible that the addressable range may change again in the
future.

In this scenario, we end up with a bogus section returned from
__section_nr (see the discussion on the thread "mm: Trigger bug on if a
section is not found in __section_nr").

Adding a check here means that we fail early and have an opportunity to
handle the error gracefully, rather than rumbling on and potentially
accessing an incorrect section.

Further discussion is also on the thread ("powerpc: Perform a bounds check
in arch_add_memory").

David said:

: I guess checking for address space wrapping would be overkill.  This
: change makes sense for architecture-independent device drivers that make
: use of the add/remove memory infrastructure (e.g., virtio-mem I am working
: on).

Link: http://lkml.kernel.org/r/20190917010752.28395-2-alastair@au1.ibm.com
Signed-off-by: Alastair D'Silva <alastair@d-silva.org>
Reviewed-by: David Hildenbrand <david@redhat.com>
Cc: Oscar Salvador <osalvador@suse.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Pavel Tatashin <pasha.tatashin@soleen.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Wei Yang <richard.weiyang@gmail.com>
Cc: Qian Cai <cai@lca.pw>
Cc: Jason Gunthorpe <jgg@ziepe.ca>
Cc: Logan Gunthorpe <logang@deltatee.com>
Cc: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>