ocfs2: move the mlog from the spinlock to outside
There is a potential task of deadlock.  Because the mask is 0, the
deadlock does not occur now.  But There is a potential task.  If someone
change the mask of mlog, but forget to modify the order of the mlog and
spin_unlock, There will be a potential deadlock.So I move the mlog from
the spinlock to outsize.

Link: http://lkml.kernel.org/r/1569813459-8601-1-git-send-email-sunny.s.zhang@oracle.com
Signed-off-by: Shuning Zhang <sunny.s.zhang@oracle.com>
Cc: Mark Fasheh <mark@fasheh.com>
Cc: Joel Becker <jlbec@evilplan.org>
Cc: Junxiao Bi <junxiao.bi@oracle.com>
Cc: Joseph Qi <jiangqi903@gmail.com>
Cc: Changwei Ge <gechangwei@live.cn>
Cc: Gang He <ghe@suse.com>
Cc: Jun Piao <piaojun@huawei.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>