mm: thp: clear PageDoubleMap flag when the last PMD map gone
File THP sets PageDoubleMap flag when the first it gets PTE mapped, but
the flag is never cleared until the THP is freed.  This result in
unbalanced state although it is not a big deal.

Clear the flag when the last compound_mapcount is gone.  It should be
cleared when all the PTE maps are gone (become PMD mapped only) as well,
but this needs check all subpage's _mapcount every time any subpage's rmap
is removed, the overhead may be not worth.  The anonymous THP also just
clears PageDoubleMap flag when the last PMD map is gone.

Link: http://lkml.kernel.org/r/1571938066-29031-1-git-send-email-yang.shi@linux.alibaba.com
Signed-off-by: Yang Shi <yang.shi@linux.alibaba.com>
Acked-by: Song Liu <songliubraving@fb.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>